moving oshdb.util.function to oshdb-util module
