+++ oshdb-util/src/main/java/org/heigit/ohsome/oshdb/util/function/OSHEntityFilter.java
@@ -1,0 +1,7 @@
+package org.heigit.ohsome.oshdb.util.function;
+
+import org.heigit.ohsome.oshdb.osh.OSHEntity;
+
+public interface OSHEntityFilter extends SerializablePredicate<OSHEntity> {
+
+}
