+++ oshdb-util/src/main/java/org/heigit/ohsome/oshdb/util/function/OSMEntityFilter.java
@@ -1,0 +1,7 @@
+package org.heigit.ohsome.oshdb.util.function;
+
+import org.heigit.ohsome.oshdb.osm.OSMEntity;
+
+public interface OSMEntityFilter extends SerializablePredicate<OSMEntity> {
+
+}
