+++ oshdb-util/src/main/java/org/heigit/ohsome/oshdb/util/function/SerializableBinaryOperator.java
