--- oshdb/src/main/java/org/heigit/ohsome/oshdb/osh/OSHEntityFilter.java
@@ -1,7 +1,0 @@
-package org.heigit.ohsome.oshdb.osh;
-
-import org.heigit.ohsome.oshdb.util.function.SerializablePredicate;
-
-public interface OSHEntityFilter extends SerializablePredicate<OSHEntity> {
-
-}
