--- oshdb/src/main/java/org/heigit/ohsome/oshdb/osm/OSMEntityFilter.java
@@ -1,7 +1,0 @@
-package org.heigit.ohsome.oshdb.osm;
-
-import org.heigit.ohsome.oshdb.util.function.SerializablePredicate;
-
-public interface OSMEntityFilter extends SerializablePredicate<OSMEntity> {
-
-}
